# src/openapi.yaml
openapi: 3.0.0
info:
  title: RPP Testbed
  description: RPP Testbed server
  version: 0.0.0
...
paths:
...
  /domains:
    head:
      operationId: domains_Check
      parameters:
        - $ref: '#/components/parameters/RequestHeaders'
      responses:
        '200':
          description: The request has succeeded.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        '400':
          description: The server could not understand the request due to invalid syntax.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse400'
        '404':
          description: The server cannot find the requested resource.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse404'
        '500':
          description: Server error
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse500'
      tags:
        - Domains
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Domain'
      security:
        - BasicAuth: []
    post:
      operationId: domains_Create
      parameters:
        - $ref: '#/components/parameters/RequestHeaders'
        - name: expect
          in: header
          required: false
          schema:
            type: string
            enum:
              - 100-continue
      responses:
        '200':
          description: The request has succeeded.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        '400':
          description: The server could not understand the request due to invalid syntax.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse400'
        '409':
          description: The request conflicts with the current state of the server.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse409'
        '500':
          description: Server error
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse500'
      tags:
        - Domains
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainCreate'
      security:
        - BasicAuth: []
  /domains/{id}:
    get:
      operationId: domains_Get
      parameters:
        - $ref: '#/components/parameters/RequestHeaders'
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        '400':
          description: The server could not understand the request due to invalid syntax.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse400'
        '404':
          description: The server cannot find the requested resource.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse404'
        '500':
          description: Server error
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse500'
      tags:
        - Domains
      security:
        - BasicAuth: []
    patch:
      operationId: domains_Update
      parameters:
        - $ref: '#/components/parameters/RequestHeaders'
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        '400':
          description: The server could not understand the request due to invalid syntax.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse400'
        '404':
          description: The server cannot find the requested resource.
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse404'
        '500':
          description: Server error
          headers:
            rpp-cltrid:
              required: false
              schema:
                type: string
            rpp-svtrid:
              required: false
              schema:
                type: string
          content:
            application/problem-json:
              schema:
                $ref: '#/components/schemas/ErrorResponse500'
      tags:
        - Domains
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainUpdate'
      security:
        - BasicAuth: []
    delete:
      operationId: domains_Delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
      tags:
        - Domains
      security:
        - BasicAuth: []
...
components:
  parameters:
    RequestHeaders:
      name: rpp-cltrid
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    ContactReference:
      type: object
      required:
        - value
        - type
      properties:
        value:
          type: string
        type:
          type: array
          items:
            $ref: '#/components/schemas/ContactType'
...
    ContactType:
      type: string
      enum:
        - registrant
        - admin
        - tech
        - billing
    DnsSec:
      type: object
      required:
        - keyTag
        - algorithm
        - digestType
        - digest
      properties:
        keyTag:
          type: string
        algorithm:
          type: string
        digestType:
          type: string
        digest:
          type: string
    Domain:
      type: object
      required:
        - name
        - authInfo
      properties:
        name:
          type: string
        status:
          type: array
          items:
            type: string
        processes:
          type: object
          properties:
            creation:
              $ref: '#/components/schemas/DomainCreation'
            renewal:
              $ref: '#/components/schemas/DomainRenewal'
            transfer:
              $ref: '#/components/schemas/DomainTransfer'
        authInfo:
          type: object
          properties:
            pw:
              type: string
            hash:
              type: string
        ns:
          $ref: '#/components/schemas/NS'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/ContactReference'
        dnsSEC:
          type: array
          items:
            $ref: '#/components/schemas/DnsSec'
        crDate:
          type: string
        exDate:
          type: string
        upDate:
          type: string
        trDate:
          type: string
        clID:
          type: string
        crID:
          type: string
      example:
        name: example.com
        processes:
          creation:
            period: P2Y
        authInfo:
          pw: password
        contacts:
          - value: registrant-id
            type:
              - registrant
    DomainCreate:
      type: object
      required:
        - name
        - authInfo
      properties:
        name:
          type: string
        processes:
          type: object
          properties:
            creation:
              $ref: '#/components/schemas/DomainCreation'
            renewal:
              $ref: '#/components/schemas/DomainRenewal'
            transfer:
              $ref: '#/components/schemas/DomainTransfer'
        authInfo:
          type: object
          properties:
            pw:
              type: string
            hash:
              type: string
        ns:
          $ref: '#/components/schemas/NS'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/ContactReference'
        dnsSEC:
          type: array
          items:
            $ref: '#/components/schemas/DnsSec'
      example:
        name: example.com
        processes:
          creation:
            period: P2Y
        authInfo:
          pw: password
        contacts:
          - value: registrant-id
            type:
              - registrant
    DomainCreation:
      type: object
      properties:
        period:
          type: string
          format: duration
...
    DomainRenewal:
      type: object
      properties:
        period:
          type: string
          format: duration
...
    DomainTransfer:
      type: object
      properties:
        period:
          type: string
          format: duration
    DomainUpdate:
      type: object
      properties:
        name:
          type: string
        processes:
          type: object
          properties:
            creation:
              $ref: '#/components/schemas/DomainCreation'
            renewal:
              $ref: '#/components/schemas/DomainRenewal'
            transfer:
              $ref: '#/components/schemas/DomainTransfer'
        authInfo:
          type: object
          properties:
            pw:
              type: string
            hash:
              type: string
        ns:
          $ref: '#/components/schemas/NSUpdate'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/ContactReference'
        dnsSEC:
          type: array
          items:
            $ref: '#/components/schemas/DnsSec'
      example:
        name: example.com
        processes:
          creation:
            period: P2Y
        authInfo:
          pw: password
        contacts:
          - value: registrant-id
            type:
              - registrant
    ErrorResponse:
      type: object
      required:
        - title
        - detail
        - type
        - status
        - instance
      properties:
        title:
          type: string
        detail:
          type: string
        type:
          type: string
        status:
          type: string
        instance:
          type: string
    ErrorResponse400:
      type: object
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse404:
      type: object
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse409:
      type: object
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse500:
      type: object
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
...
    HostInfoAttr:
      type: object
      properties:
        addr:
          type: object
          properties:
            ipv4:
              type: array
              items:
                type: string
            ipv6:
              type: array
              items:
                type: string
      allOf:
        - $ref: '#/components/schemas/HostInfoObj'
...
    HostInfoObj:
      type: object
      required:
        - name
      properties:
        name:
          type: string
...
    NS:
      oneOf:
        - $ref: '#/components/schemas/NSHostObj'
        - $ref: '#/components/schemas/NSHostAttr'
    NSHostAttr:
      type: object
      required:
        - hostAttr
      properties:
        hostAttr:
          type: array
          items:
            $ref: '#/components/schemas/HostInfoAttr'
    NSHostAttrUpdate:
      type: object
      properties:
        hostAttr:
          type: array
          items:
            $ref: '#/components/schemas/HostInfoAttr'
    NSHostObj:
      type: object
      required:
        - hostObj
      properties:
        hostObj:
          type: array
          items:
            $ref: '#/components/schemas/HostInfoObj'
    NSHostObjUpdate:
      type: object
      properties:
        hostObj:
          type: array
          items:
            $ref: '#/components/schemas/HostInfoObj'
    NSUpdate:
      oneOf:
        - $ref: '#/components/schemas/NSHostObjUpdate'
        - $ref: '#/components/schemas/NSHostAttrUpdate'
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: Basic
servers:
  - url: https://localhost:8000
    description: Local RPP Testbed server
    variables: {}